rev 20
found out the hard way that when you pass in a selector boolean you run into where you can call it up in a template
when you define a boolean input in the blueprint itself using selector: boolean, that’s not an entity, it’s just a true/false value in the blueprint context. It doesn’t exist as an entity in HA. That means you cannot reliably reference it directly in actions or debug notifications in the same way you do with !input entities — unless you assign it to a variable first.

Thus I did need to assign the input into a variable and then use that for the templates.

also removed the in_ for the bool due to it not being an input_boolean